requirements.txt
Django==5.2.1
django-crispy-forms==2.4
crispy-bootstrap5==2025.4
Pillow==11.2.1
python-decouple==3.8
django-tables2==2.7.5
django-filter==25.1
reportlab==4.4.0
XlsxWriter==3.2.3 